Refactoring new ui create filter
