+++ app/src/main/java/com/github/damontecres/stashapp/ui/components/filter/CreateFilter.kt
@@ -3,15 +3,16 @@ package com.github.damontecres.stashapp.ui.components.filter
 import androidx.activity.compose.BackHandler
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.background
-import androidx.compose.foundation.focusGroup
 import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.interaction.collectIsFocusedAsState
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.widthIn
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
@@ -30,7 +31,6 @@ import androidx.compose.ui.draw.blur
 import androidx.compose.ui.focus.FocusRequester
 import androidx.compose.ui.focus.focusProperties
 import androidx.compose.ui.focus.focusRequester
-import androidx.compose.ui.focus.focusRestorer
 import androidx.compose.ui.graphics.graphicsLayer
 import androidx.compose.ui.input.key.Key
 import androidx.compose.ui.input.key.KeyEventType
@@ -48,7 +48,6 @@ import androidx.tv.material3.Text
 import com.apollographql.apollo.api.Optional
 import com.github.damontecres.stashapp.R
 import com.github.damontecres.stashapp.api.fragment.StashData
-import com.github.damontecres.stashapp.api.type.CriterionModifier
 import com.github.damontecres.stashapp.api.type.StashDataFilter
 import com.github.damontecres.stashapp.data.DataType
 import com.github.damontecres.stashapp.data.StashFindFilter
@@ -56,7 +55,6 @@ import com.github.damontecres.stashapp.data.flip
 import com.github.damontecres.stashapp.filter.CreateFilterViewModel
 import com.github.damontecres.stashapp.filter.FilterOption
 import com.github.damontecres.stashapp.filter.filterSummary
-import com.github.damontecres.stashapp.filter.findFilterSummary
 import com.github.damontecres.stashapp.filter.getFilterOptions
 import com.github.damontecres.stashapp.navigation.Destination
 import com.github.damontecres.stashapp.navigation.NavigationManager
@@ -66,9 +64,7 @@ import com.github.damontecres.stashapp.ui.components.CircularProgress
 import com.github.damontecres.stashapp.ui.tryRequestFocus
 import com.github.damontecres.stashapp.ui.util.ifElse
 import com.github.damontecres.stashapp.util.StashCoroutineExceptionHandler
-import com.github.damontecres.stashapp.util.isNotNullOrBlank
 import kotlinx.coroutines.launch
-import java.util.Date
 import kotlin.reflect.full.createInstance
 
 internal const val TAG = "CreateFilter"
@@ -239,9 +235,19 @@ fun CreateFilterColumns(
                 visible = !findFilterFocused && !objectFilterFocused,
             ) {
                 LazyColumn(
-                    modifier = Modifier,
+                    modifier =
+                        Modifier
+                            .widthIn(min = listWidth)
+//                        .width(listWidth)
+                            .background(
+                                color = MaterialTheme.colorScheme.secondaryContainer.copy(alpha = .75f),
+                                shape = RoundedCornerShape(16.dp),
+                            ),
+                    verticalArrangement = Arrangement.spacedBy(8.dp),
+                    horizontalAlignment = Alignment.Start,
+                    contentPadding = PaddingValues(8.dp),
                 ) {
-                    item {
+                    stickyHeader {
                         val str =
                             if (filterSummaries.isEmpty()) {
                                 R.string.no_filters_set
@@ -255,8 +261,7 @@ fun CreateFilterColumns(
                             textAlign = TextAlign.Center,
                             modifier =
                                 Modifier
-                                    .fillMaxWidth()
-                                    .padding(8.dp),
+                                    .fillParentMaxWidth(),
                         )
                     }
                     items(filterSummaries) {
@@ -267,8 +272,7 @@ fun CreateFilterColumns(
                             textAlign = TextAlign.Start,
                             modifier =
                                 Modifier
-                                    .fillMaxWidth()
-                                    .padding(8.dp),
+                                    .fillMaxWidth(),
                         )
                     }
                 }
@@ -544,134 +548,5 @@ fun CreateFilterColumns(
             onSave = it.onSave,
             onDismiss = { inputDurationAction = null },
         )
-    }
-}
-
-data class InputTextAction(
-    val title: String,
-    val value: String?,
-    val keyboardType: KeyboardType,
-    val onSubmit: (String) -> Unit,
-)
-
-data class InputCriterionModifier(
-    val filterName: String,
-    val allowedModifiers: List<CriterionModifier>,
-    val onClick: (CriterionModifier) -> Unit,
-)
-
-data class SelectFromListAction(
-    val filterName: String,
-    val options: List<String>,
-    val currentOptions: List<String>,
-    val multiSelect: Boolean,
-    val onSubmit: (indices: List<Int>) -> Unit,
-)
-
-data class MultiCriterionInfo(
-    val name: String,
-    val dataType: DataType,
-    val initialValues: List<IdName>,
-    val onAdd: (IdName) -> Unit,
-    val onSave: (List<IdName>) -> Unit,
-)
-
-data class InputDateAction(
-    val name: String,
-    val value: Date,
-    val onSave: (Date) -> Unit,
-)
-
-data class InputDurationAction(
-    val name: String,
-    val value: Int?,
-    val onSave: (Int) -> Unit,
-)
-
-@Composable
-fun BasicFilterSettings(
-    dataType: DataType,
-    name: String?,
-    findFilter: StashFindFilter,
-    objectFilter: StashDataFilter,
-    objectFilterCount: Int,
-    resultCount: Int,
-    onFilterNameClick: () -> Unit,
-    findFilterOnClick: () -> Unit,
-    objectFilterOnClick: () -> Unit,
-    findFilterInteractionSource: MutableInteractionSource,
-    objectFilterInteractionSource: MutableInteractionSource,
-    onSubmit: (Boolean) -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    val context = LocalContext.current
-    val focusRequester = remember { FocusRequester() }
-    LazyColumn(
-        modifier =
-            modifier
-                .focusGroup()
-                .focusRestorer(focusRequester),
-        verticalArrangement = Arrangement.spacedBy(0.dp),
-    ) {
-        item {
-            SimpleListItem(
-                title = stringResource(R.string.stashapp_filter_name),
-                subtitle = name,
-                showArrow = false,
-                onClick = onFilterNameClick,
-                modifier = Modifier.focusRequester(focusRequester),
-            )
-        }
-        item {
-            SimpleListItem(
-                title = stringResource(R.string.sort_by),
-                subtitle = findFilterSummary(context, dataType, findFilter),
-                showArrow = true,
-                onClick = findFilterOnClick,
-                interactionSource = findFilterInteractionSource,
-            )
-        }
-        item {
-            SimpleListItem(
-                title = stringResource(R.string.stashapp_filters),
-                subtitle =
-                    when (objectFilterCount) {
-                        // TODO i18n for plurals
-                        1 -> "1 " + stringResource(R.string.stashapp_filter)
-                        in 2..Int.MAX_VALUE -> objectFilterCount.toString() + " " + stringResource(R.string.stashapp_filters)
-                        else -> null
-                    },
-                showArrow = true,
-                onClick = objectFilterOnClick,
-                interactionSource = objectFilterInteractionSource,
-            )
-        }
-        item {
-            SimpleListItem(
-                title = stringResource(R.string.submit_without_saving),
-                subtitle =
-                    if (resultCount >= 0) {
-                        resultCount.toString() + " " + stringResource(R.string.results)
-                    } else {
-                        null
-                    },
-                showArrow = false,
-                onClick = { onSubmit.invoke(false) },
-            )
-        }
-        item {
-            SimpleListItem(
-                enabled = name.isNotNullOrBlank(),
-                title = stringResource(R.string.save_and_submit),
-                subtitle =
-                    if (name.isNotNullOrBlank()) {
-                        null
-                    } else {
-                        stringResource(R.string.save_and_submit_no_name_desc)
-                    },
-                showArrow = false,
-                onClick = { onSubmit.invoke(true) },
-            )
-        }
     }
 }
